Restructure callflow.py -> main.py; removes dataset.py -> supergraph.py
