+++ callflow/__init__.py
@@ -5,12 +5,9 @@ from .timer import Timer
 
 from .datastructures.graphframe import GraphFrame
 
-from .datastructures.dataset import Dataset
-from .datastructures.ensemble_graph import EnsembleGraph
-from .datastructures.supergraph_base import SuperGraph
+from .datastructures.supergraph import SuperGraph
+from .datastructures.ensemblegraph import EnsembleGraph
 from .datastructures.cct import CCT
 
 
-from .callflow_base import BaseCallFlow
-from .callflow_single import SingleCallFlow
-from .callflow_ensemble import EnsembleCallFlow
+from .main import CallFlow
