+++ callflow/datastructures/supergraph_ensemble.py
@@ -89,106 +89,6 @@ class EnsembleSuperGraph(SuperGraph):
             self.add_edge_attributes()
         print(self.timer)
 
-    def create_target_maps(self):
-        self.target_df = {}
-        self.target_modules = {}
-        self.target_module_group_df = {}
-        self.target_module_name_group_df = {}
-        self.target_module_callsite_map = {}
-        self.target_module_time_inc_map = {}
-        self.target_module_time_exc_map = {}
-        self.target_name_time_inc_map = {}
-        self.target_name_time_exc_map = {}
-
-        for run in self.runs:
-            # Reduce the entire_df to respective target dfs.
-            self.target_df[run] = self.entire_df.loc[self.entire_df["dataset"] == run]
-
-            # Unique modules in the target run
-            self.target_modules[run] = self.target_df[run]["module"].unique()
-
-            # Group the dataframe in two ways.
-            # 1. by module
-            # 2. by module and callsite
-            self.target_module_group_df[run] = self.target_df[run].groupby(["module"])
-            self.target_module_name_group_df[run] = self.target_df[run].groupby(
-                ["module", "name"]
-            )
-
-            # Module map for target run {'module': [Array of callsites]}
-            self.target_module_callsite_map[run] = (
-                self.target_module_group_df[run]["name"].unique().to_dict()
-            )
-
-            # Inclusive time maps for the module level and callsite level.
-            self.target_module_time_inc_map[run] = (
-                self.target_module_group_df[run]["time (inc)"].max().to_dict()
-            )
-            self.target_name_time_inc_map[run] = (
-                self.target_module_name_group_df[run]["time (inc)"].max().to_dict()
-            )
-
-            # Exclusive time maps for the module level and callsite level.
-            self.target_module_time_exc_map[run] = (
-                self.target_module_group_df[run]["time"].max().to_dict()
-            )
-            self.target_name_time_exc_map[run] = (
-                self.target_module_name_group_df[run]["time"].max().to_dict()
-            )
-
-    def create_ensemble_maps(self):
-        self.modules = self.entire_df["module"].unique()
-
-        self.module_name_group_df = self.entire_df.groupby(["module", "name"])
-        self.module_group_df = self.entire_df.groupby(["module"])
-        self.name_group_df = self.entire_df.groupby(["name"])
-
-        # Module map for ensemble {'module': [Array of callsites]}
-        self.module_callsite_map = self.module_group_df["name"].unique().to_dict()
-
-        # Inclusive time maps for the module level and callsite level.
-        self.module_time_inc_map = self.module_group_df["time (inc)"].max().to_dict()
-        self.name_time_inc_map = self.module_name_group_df["time (inc)"].max().to_dict()
-
-        # Exclusive time maps for the module level and callsite level.
-        self.module_time_exc_map = self.module_group_df["time"].max().to_dict()
-        self.name_time_exc_map = self.module_name_group_df["time"].max().to_dict()
-
-    def construct_cycle_free_paths(self, path):
-        ret = []
-        moduleMapper = {}
-        dataMap = {}
-
-        if isinstance(path, float):
-            return []
-        path = make_list(path)
-        for idx, elem in enumerate(path):
-            callsite = elem.split("=")[1]
-            module = elem.split("=")[0]
-            if module not in dataMap:
-                moduleMapper[module] = 0
-                dataMap[module] = [
-                    {"callsite": callsite, "module": module, "level": idx}
-                ]
-            else:
-                flag = [p["level"] == idx for p in dataMap[module]]
-                if np.any(np.array(flag)):
-                    moduleMapper[module] += 1
-                    dataMap[module].append(
-                        {
-                            "callsite": callsite,
-                            "module": module + "=" + callsite,
-                            "level": idx,
-                        }
-                    )
-                else:
-                    dataMap[module].append(
-                        {"callsite": callsite, "module": module, "level": idx}
-                    )
-            ret.append(dataMap[module][-1])
-
-        return ret
-
     def create_source_targets(self, component_path):
         module = ""
         edges = []
